refactor(platform-github): 公共工具文件改名 utils，review 态映射改常量表
mappers.ts → utils.ts（承载跨领域公共工具方法 mapUser）。PR 领域内 review 决断态到活动
事件类型 / reviewer 状态的嵌套三元 / if-else 改为常量 Record 映射表（ACTIVITY_KIND_BY_STATE
/ REVIEWER_STATUS_BY_STATE），更易读、易维护。行为等价。
